Trait minimint::net::connect::Connector

pub trait Connector<M> {
    fn connect_framed<'life0, 'async_trait>(
        &'life0 self,
        destination: String,
        peer: PeerId
    ) -> Pin<Box<dyn Future<Output = ConnectResult<M>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn listen<'life0, 'async_trait>(
        &'life0 self,
        bind_addr: String
    ) -> Pin<Box<dyn Future<Output = Result<ConnectionListener<M>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn to_any(self) -> AnyConnector<M>
    where
        Self: Sized + Send + Sync + Unpin + 'static,
    { ... }
}

Allows to connect to peers and to listen for incoming connections

Connections are message based (FramedTransport) and should be authenticated and encrypted for production deployments.

Required Methods

fn connect_framed<'life0, 'async_trait>(
&'life0 self,
destination: String,
peer: PeerId
) -> Pin<Box<dyn Future<Output = ConnectResult<M>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Connect to a destination

fn listen<'life0, 'async_trait>(
&'life0 self,
bind_addr: String
) -> Pin<Box<dyn Future<Output = Result<ConnectionListener<M>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Listen for incoming connections on bind_addr

Provided Methods

fn to_any(self) -> AnyConnector<M> where
Self: Sized + Send + Sync + Unpin + 'static,

Transform this concrete Connector into an owned trait object version of itself