Type Definition minimint::net::framed::TcpBidiFramed

pub type TcpBidiFramed<T> = BidiFramed<T, OwnedWriteHalf, OwnedReadHalf>;

Special case for tokio TcpStream based BidiFramed instances

Implementations

impl<T> TcpBidiFramed<T> where
T: Serialize + DeserializeOwned,

pub fn new_from_tcp(stream: TcpStream) -> TcpBidiFramed<T>

Special constructor for tokio TCP connections.

Tokio TcpStream implements an efficient method of splitting the stream into a read and a write half this constructor takes advantage of.