Module minimint::net::peers

Implements a connection manager for communication with other federation members

The main interface is PeerConnections and its main implementation is ReconnectPeerConnections, see these for details.

Structs

CommonPeerConnectionState 🔒

ConnectedPeerConnectionState 🔒

ConnectionConfig

Information needed to connect to one other federation member

DisconnectedPeerConnectionState 🔒

NetworkConfig

Specifies the network configuration for federation-internal communication

PeerConnection 🔒

PeerConnectionStateMachine 🔒

PeerMessage

Internal message type for ReconnectPeerConnections, just public because it appears in the public interface.

ReconnectPeerConnections

Connection manager that automatically reconnects to peers

Enums

PeerConnectionState 🔒

Constants

MAX_FAIL_RECONNECT_COUNTER 🔒

Maximum connection failures we consider for our back-off strategy

Traits

PeerConnections

Connection manager that tries to keep connections open to all peers

Type Definitions

AnyPeerConnections

Owned PeerConnections trait object type

PeerConnector

Owned Connector trait object used by ReconnectPeerConnections