Trait minimint::rng::RngGenerator

pub trait RngGenerator: Sync + Send {
    type Rng: RngCore + CryptoRng;

    fn get_rng(&self) -> Self::Rng;
}

Cheaply generates a new random number generator. Since these need to be generated often to avoid locking them when used by different threads the construction should be rather cheap.

Required Associated Types

type Rng: RngCore + CryptoRng

Required Methods

fn get_rng(&self) -> Self::Rng